Created useAuth hook
